pyproject.toml
[tool.poetry]
name = "nest-doorbell-server"
version = "0.1.0"
description = "Google Nest Doorbell notification server"
authors = ["AlexPerrotquet <alex@perroquet.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
google-cloud-pubsub = "^2.18.4"
google-cloud-storage = "^2.13.0"
requests = "^2.31.0"
aiofiles = "^23.2.1"
python-multipart = "^0.0.6"
pydantic = "2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
